commit homework 03 follow comment
