+++ BaiTap11/DuongNT/Protocol/Protocol/Controllers/Ex3/Ex3ViewController.swift
@@ -27,6 +27,10 @@ class Ex3ViewController: BaseViewController {
         super.didReceiveMemoryWarning()
     }
 
+    deinit {
+        NotificationCenter.default.removeObserver(self, name: NSNotification.Name("REGIONNOTIFICATION"), object: nil)
+    }
+
     // MARK: - Config
     override func setupUI() {
         super.setupUI()
