+++ BaiTap11/DuongNT/Protocol/Protocol/Controllers/Ex3/Huyen/HuyenViewController.swift
@@ -1,5 +1,5 @@
 //
-//  Mien1ViewController.swift
+//  Tinh1ViewController.swift
 //  Protocol
 //
 //  Created by Nguyen Duong on 8/12/19.
@@ -8,29 +8,31 @@
 
 import UIKit
 
-class Mien1ViewController: BaseViewController {
+class HuyenViewController: BaseViewController {
 
     @IBOutlet weak var tableView: UITableView!
+
     // MARK: - Properties
     var mien: Mien?
-    var tinhList: [Tinh] = []
+    var tinh: Tinh?
+    var huyenList: [Huyen] = []
 
     // MARK: - Config
     override func setupUI() {
         super.setupUI()
-        self.title = mien?.name
+        self.title = tinh?.name
         tableView.register(UINib(nibName: "MienCell", bundle: nil), forCellReuseIdentifier: "cell")
     }
 
     override func setupData() {
         super.setupData()
-        if let mien = mien {
-            tinhList = mien.tinh
+        if let tinh = tinh {
+            huyenList = tinh.huyen
         }
     }
 }
 
-extension Mien1ViewController: UITableViewDelegate, UITableViewDataSource {
+extension HuyenViewController: UITableViewDelegate, UITableViewDataSource {
 
     // MARK: - Table view data source
     func numberOfSections(in tableView: UITableView) -> Int {
@@ -38,8 +40,8 @@ extension Mien1ViewController: UITableViewDelegate, UITableViewDataSource {
     }
 
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return tinhList.count
-
+        return huyenList.count
+        
     }
 
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
@@ -48,29 +50,24 @@ extension Mien1ViewController: UITableViewDelegate, UITableViewDataSource {
 
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "cell") as? MienCell
-        let tinh = tinhList[indexPath.row]
-        cell?.nameLabel.text = tinh.name
+        let huyen = huyenList[indexPath.row]
+        cell?.nameLabel.text = huyen.name
         return cell!
     }
 
     // MARK: - Table view delegate
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         tableView.deselectRow(at: indexPath, animated: true)
-        switch indexPath.row {
-        case 0:
-            let vc = Tinh1ViewController()
-            vc.mien = mien
-            vc.tinh = self.tinhList[indexPath.row]
-            self.navigationController?.pushViewController(vc, animated: true)
-        default:
-            // create the alert
-            let alert = UIAlertController(title: "Warning",
-                                          message: "Please, create new View Controller.",
-                                          preferredStyle: UIAlertController.Style.alert)
-            // add an action (button)
-            alert.addAction(UIAlertAction(title: "OK", style: UIAlertAction.Style.default, handler: nil))
-            // show the alert
-            self.present(alert, animated: true, completion: nil)
+        var dict: Dictionary<String, AnyObject> = Dictionary()
+        dict.updateValue(self.huyenList[indexPath.row].name as AnyObject, forKey: "huyen")
+        dict.updateValue(self.tinh?.name as AnyObject, forKey: "tinh")
+        dict.updateValue(self.mien?.name as AnyObject, forKey: "mien")
+        NotificationCenter.default.post(name: NSNotification.Name("REGIONNOTIFICATION"), object: nil, userInfo: dict)
+        for controller in self.navigationController!.viewControllers as Array {
+            if controller.isKind(of: Ex3ViewController.self) {
+                self.navigationController!.popToViewController(controller, animated: true)
+                break
+            }
         }
     }
 }
