--- BaiTap11/DuongNT/Protocol/Protocol/Controllers/Ex3/Mien2/Mien2ViewController.swift
@@ -1,35 +1,0 @@
-//
-//  Mien2ViewController.swift
-//  Protocol
-//
-//  Created by Nguyen Duong on 8/12/19.
-//  Copyright © 2019 Tien Le P. All rights reserved.
-//
-
-import UIKit
-
-class Mien2ViewController: BaseViewController {
-
-    // MARK: - Properties
-    var mien: Mien?
-
-
-    // MARK: - Life Cicle
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-    }
-
-    // MARK: - Config
-    override func setupUI() {
-        super.setupUI()
-        self.title = mien?.name
-    }
-
-    override func setupData() {
-        super.setupData()
-    }
-}
